+++ lib/getch/filesystem/zfs/config.rb
@@ -23,16 +23,17 @@ module Getch
             'title Gentoo Linux',
             'linux /vmlinuz',
             'initrd /initramfs',
-            "options resume=UUID=#{@uuid_swap} root=#{@pool_name}/ROOT/gentoo init=#{@init} dozfs rw"
+            "options resume=UUID=#{@uuid_swap} root=ZFS=#{@pool_name}/ROOT/gentoo init=#{@init} dozfs"
           ]
           File.write("#{dir}/gentoo.conf", datas_gentoo.join("\n"))
         end
 
+        # See https://wiki.gentoo.org/wiki/ZFS#ZFS_root
         def grub
           return if Helpers::efi?
           file = "#{@root_dir}/etc/default/grub"
           cmdline = [ 
-            "GRUB_CMDLINE_LINUX=\"resume=UUID=#{@uuid_swap} root=#{@pool_name}/ROOT/gentoo init=#{@init} dozfs rw\""
+            "GRUB_CMDLINE_LINUX=\"resume=UUID=#{@uuid_swap} root=ZFS=#{@pool_name}/ROOT/gentoo init=#{@init} dozfs\""
           ]
           File.write("#{file}", cmdline.join("\n"), mode: 'a')
         end
@@ -40,16 +41,14 @@ module Getch
         private
 
         def gen_uuid
-          @uuid_swap = `lsblk -o "UUID" #{@lv_swap} | tail -1`.chomp() if @dev_swap
-          @uuid_root = `lsblk -o "UUID" #{@lv_root} | tail -1`.chomp() if @dev_root
-          @uuid_dev_root = `lsblk -o "UUID" #{@dev_root} | tail -1`.chomp() if @dev_root
+          @uuid_swap = `lsblk -o "UUID" #{@dev_swap} | tail -1`.chomp()
           @uuid_boot = `lsblk -o "UUID" #{@dev_boot} | tail -1`.chomp() if @dev_boot
           @uuid_boot_efi = `lsblk -o "UUID" #{@dev_boot_efi} | tail -1`.chomp() if @dev_boot_efi
         end
 
         def data_fstab
           boot_efi = @dev_boot_efi ? "UUID=#{@uuid_boot_efi} /boot/efi vfat noauto,noatime 1 2" : ''
-          swap = @lv_swap ? "UUID=#{@uuid_swap} none swap discard 0 0" : ''
+          swap = @dev_swap ? "UUID=#{@uuid_swap} none swap discard 0 0" : ''
 
           [ boot_efi, swap ]
         end