+++ lib/getch/gentoo/boot.rb
@@ -13,6 +13,7 @@ module Getch
         @config.fstab
         bootloader
         password
+        cleanup
         the_end
       end
 
@@ -75,6 +76,10 @@ module Getch
       end
 
       private
+
+      def cleanup
+        Getch::Emerge.new("emerge --depclean").run!
+      end
 
       def the_end
         #Helpers::exec_or_die("umount -l /mnt/gentoo/dev{/shm,/pts,}")